refactor: remove QueryAnalysis
