+++ apollo-router/src/services/query_parsing/cache.rs
@@ -6,7 +6,7 @@ use lru::LruCache;
 use tokio::sync::Mutex;
 
 use crate::compute_job::MaybeBackPressureError;
-use crate::services::layers::query_analysis::ParsedDocument;
+use crate::services::query_parsing::ParsedDocument;
 use crate::services::query_parsing::Request;
 use crate::services::query_parsing::ServiceError;
 use crate::spec::SpecError;
